% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{log_modified}
\alias{log_modified}
\title{Modified log function (internal function)}
\usage{
log_modified(x)
}
\arguments{
\item{x}{A \code{numeric} value assumed such that x >= 0.}
}
\value{
Modified log function
}
\description{
Simple function that return log(x) if x > 0 and 0 otherwise.
}
\author{
Stephane Guerrier
}
