% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{marginal_mle}
\alias{marginal_mle}
\title{Compute (marginalized) MLE based on the partial information R1 and R3.}
\usage{
marginal_mle(
  R1,
  R3,
  n,
  pi0,
  gamma = 0.05,
  alpha = 0,
  beta = 0,
  alpha0 = 0,
  beta0 = 0,
  ...
)
}
\arguments{
\item{R1}{A \code{numeric} that provides the number of participants in the survey sample that were tested positive with both (medical) testing devices (and are, thus, members of the sub-population).}

\item{R3}{A \code{numeric} that provides the number of participants in the survey sample that are tested positive only with the second testing device.}

\item{n}{A \code{numeric} that provides the sample size.}

\item{pi0}{A \code{numeric} that provides the prevalence or proportion of people (in the whole population) who are positive, as measured through a non-random,
but systematic sampling (e.g. based on medical selection).}

\item{gamma}{A \code{numeric} that used to compute a (1 - gamma) confidence region for the proportion. Default value is \code{0.05}.}

\item{alpha}{A \code{numeric} that provides the False Negative (FN) rate for the sample R. Default value is \code{0}.}

\item{beta}{A \code{numeric} that provides the False Positive (FP) rate for the sample R. Default value is \code{0}.}

\item{alpha0}{A \code{numeric} that provides the False Negative (FN) rate for the sample R0. Default value is \code{0}.}

\item{beta0}{A \code{numeric} that provides the False Positive (FP) rate for the sample R0. Default value is \code{0}.}

\item{...}{Additional arguments.}
}
\value{
A \code{cpreval} object with the structure:
\itemize{
 \item estimate:    Estimated proportion.
 \item sd:          Estimated standard error of the estimator.
 \item ci_asym:     Asymptotic confidence interval at the 1 - gamma confidence level.
 \item gamma:       Confidence level (i.e. 1 - gamma) for confidence intervals.
 \item method:      Estimation method (in this case marginal mle)
 \item measurement: A vector with (alpha0, alpha, beta0, beta)
 \item ...:         Additional parameters
}
}
\description{
Proportion estimated using the MLE and confidence intervals based the asymptotic distribution of the estimator.
}
\examples{
# Samples without measurement error
X = sim_Rs(theta = 3/100, pi0 = 1/100, n = 1500, seed = 18)
mle(R1 = X$R1, R3 = X$R3, n = X$n, pi0 = X$pi0)

# With measurement error
X = sim_Rs(theta = 30/1000, pi0 = 10/1000, n = 1500, alpha0 = 0.01,
alpha = 0.01, beta0 = 0.05, beta = 0.05, seed = 18)
marginal_mle(R1 = X$R1, R3 = X$R3, n = X$n, pi0 = X$pi0)
marginal_mle(R1 = X$R1, R3 = X$R3, n = X$n, pi0 = X$pi0,
alpha0 = 0.01, alpha = 0.01, beta0 = 0.05, beta = 0.05)
}
\author{
Stephane Guerrier
}
