% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{neg_log_lik}
\alias{neg_log_lik}
\title{Negative Log-Likelihood function}
\usage{
neg_log_lik(theta, Rvect, n, pi0, alpha, beta, alpha0, beta0)
}
\arguments{
\item{theta}{A \code{numeric} value for the paramerer of interest.}

\item{Rvect}{A \code{vector} of observations, i.e. (R1, R2, R3, R4).}

\item{n}{A \code{numeric} that provides the sample size.}

\item{alpha}{A \code{numeric} that provides the False Negative (FN) rate for the sample R. Default value is \code{0}.}

\item{beta}{A \code{numeric} that provides the False Positive (FP) rate for the sample R. Default value is \code{0}.}

\item{alpha0}{A \code{numeric} that provides the False Negative (FN) rate for the sample R0. Default value is \code{0}.}

\item{beta0}{A \code{numeric} that provides the False Positive (FP) rate for the sample R0. Default value is \code{0}.}

\item{...}{Additional arguments.}
}
\value{
Negative log-likelihood.
}
\description{
Log-Likelihood function based on R1, R2, R3 and R4 multiplied by -1.
}
\author{
Stephane Guerrier
}
