% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{print.cpreval}
\alias{print.cpreval}
\title{Print estimation results}
\usage{
\method{print}{cpreval}(x, ...)
}
\arguments{
\item{x}{A \code{cpreval} object}

\item{...}{Further arguments passed to or from other methods}
}
\value{
Prints object
}
\description{
Simple print function for the four estimators (i.e. mle, conditional mle, moment based and survey sample).
}
\examples{
X = sim_Rs(p = 3/100, pi0 = 1/100, n = 1500, seed = 18)
survey_sample(X$R, X$n)
}
\author{
Stephane Guerrier
}
\keyword{internal}
