% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{print.cpreval_sim}
\alias{print.cpreval_sim}
\title{Print (simulated) sample}
\usage{
\method{print}{cpreval_sim}(x, ...)
}
\arguments{
\item{x}{A \code{cpreval_sim} object}

\item{...}{Further arguments passed to or from other methods}
}
\value{
Prints object
}
\description{
Simple print function for the \code{cpreval_sim} objects
}
\examples{
# Samples without measurement error
sim_Rs(theta = 3/100, pi0 = 1/100, n = 1500, seed = 18)

# With measurement error
sim_Rs(theta = 3/100, pi0 = 1/100, n = 1500, alpha0 = 0,
alpha = 0.01, beta = 0.05, seed = 18)
}
\author{
Stephane Guerrier
}
\keyword{internal}
